Escape attributes even if no whitelist is configured for an element
The logic for detecting that an attribute was not allowed consisted in:

    allowedAttrs !== undefined && allowedAttrs.indexOf(attr) === -1 && !wl.ATTRIBUTES['*'].has(attr)

For elements for which no `ATTRIBUTES` were configured, such as `<details>`,
variable `allowedAttrs` would be undefined and cause the entire check to
fail, mistakenly allowing any attributes, such as `<details ontoggle=hack()>`.

This commit fixes the logic, and splits the condition into named boolean
variables to make it easier to verify.

It also includes `<details ontoggle=f()></details>` as a unit test.